requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-mineru>=0.0.1
langchain-openai>=0.1.0
langchain-qdrant>=0.1.0
qdrant-client>=1.7.0
tenacity>=8.2.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
